changed int to uint64_t in render_sprite_manager function

+++ lib_dev/src/sprite/SDLite_sprite_manager.c
@@ -121,7 +121,7 @@ int SDLite_Render_SpriteManager(SDLite_Display *display, SDLite_Sprite_Manager *
 
         SDLite_Sprite **sprites = layer->sprites;
 
-        for(int a = 0; a < layer->amount_sprites; a++){
+        for(uint64_t a = 0; a < layer->amount_sprites; a++){
 
             result |= (bool)SDLite_Render_Sprite(display, sprites[a]);
